# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Serverless function execution platform (FaaS) API",
        "title": "SoftGate API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/functions": {
            "get": {
                "description": "Get a list of all registered functions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "List all functions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FunctionListItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new serverless function with code and parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Create a new function",
                "parameters": [
                    {
                        "description": "Function to create",
                        "name": "function",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateFunctionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Function"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/functions/{id}": {
            "get": {
                "description": "Get detailed information about a specific function",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Get function details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Function ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Function"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a function and remove its stored code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Delete a function",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Function ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/functions/{id}/invocations": {
            "get": {
                "description": "Get execution history for a function",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "List function invocations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Function ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InvocationListItem"
                            }
                        }
                    }
                }
            }
        },
        "/functions/{id}/invocations/{invocationId}": {
            "get": {
                "description": "Poll for the result of a function invocation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Get invocation result",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Function ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Invocation ID",
                        "name": "invocationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InvokeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/functions/{id}/invoke": {
            "post": {
                "description": "Execute a function with given parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Invoke a function",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Function ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Input parameters",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InvokeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InvokeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateFunctionRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FunctionParam"
                    }
                },
                "runtime": {
                    "type": "string"
                },
                "sample_event": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "models.Function": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "code_s3_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FunctionParam"
                    }
                },
                "runtime": {
                    "type": "string"
                },
                "sample_event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.FunctionListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "runtime": {
                    "type": "string"
                }
            }
        },
        "models.FunctionParam": {
            "type": "object",
            "properties": {
                "default_value": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "function_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.InvocationListItem": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "function_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "input_event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "invoked_at": {
                    "type": "string"
                },
                "output_result": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.InvokeRequest": {
            "type": "object",
            "properties": {
                "params": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "models.InvokeResponse": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "function_id": {
                    "type": "integer"
                },
                "input_event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "invocation_id": {
                    "type": "integer"
                },
                "logged_at": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}